Sana/requirements.txt
git+https://github.com/huggingface/diffusers.git@v0.32.0
git+https://github.com/openai/CLIP.git
accelerate==1.2.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
autoroot==1.0.1
autorootcwd==1.0.1
beautifulsoup4==4.12.3
braceexpand==0.1.7
bs4==0.0.2
came-pytorch==0.1.3
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
gitdb==4.0.11
gitpython==3.1.43
gradio==5.9.0
gradio-client==1.5.2
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
identify==2.6.3
idna==3.10
image-reward==1.5
importlib-metadata==8.5.0
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
jinja2==3.1.3
kiwisolver==1.4.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==1.7.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
optimum==1.23.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
patch-conv==0.0.1b0
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pre-commit==4.0.1
prompt-toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pydantic==2.10.3
pydantic-core==2.27.1
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
pyrallis==0.3.1
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytorch-fid==0.3.0
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.3
safehttpx==0.1.6
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spaces==0.31.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboardx==2.6.2.2
termcolor==2.5.0
timm==0.6.13
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.0 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.0.0
typer==0.15.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
wandb==0.19.1
wcwidth==0.2.13
webdataset==0.2.100
websockets==14.1
xformers==0.0.27.post2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
